Change name to namespace_state.
