+++ metricbeat/include/list_docker.go
@@ -37,6 +37,7 @@ import (
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/container"
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/controllermanager"
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/event"
+	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/namespace_state"
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/node"
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/pod"
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/proxy"
@@ -46,7 +47,6 @@ import (
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/state_daemonset"
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/state_deployment"
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/state_job"
-	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/state_namespace"
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/state_node"
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/state_persistentvolume"
 	_ "github.com/elastic/beats/v7/metricbeat/module/kubernetes/state_persistentvolumeclaim"
