+++ metricbeat/module/kubernetes/namespace_state/_meta/data.json
