+++ metricbeat/module/kubernetes/namespace_state/_meta/fields.yml
