+++ metricbeat/module/kubernetes/namespace_state/_meta/test/ksm.v2.7.0.plain.expected
