+++ metricbeat/module/kubernetes/namespace_state/_meta/test/ksm.v2.8.2.plain.expected
