+++ metricbeat/module/kubernetes/namespace_state/_meta/test/ksm.v2.9.2.plain.expected
