+++ metricbeat/module/kubernetes/namespace_state/_meta/testdata/docs.plain-expected.json
