+++ metricbeat/module/kubernetes/namespace_state/_meta/testdata/ksm.v2.7.0.plain-expected.json
