+++ metricbeat/module/kubernetes/namespace_state/_meta/testdata/ksm.v2.8.2.plain-expected.json
