+++ metricbeat/module/kubernetes/namespace_state/_meta/testdata/ksm.v2.9.2.plain-expected.json
