+++ metricbeat/module/kubernetes/namespace_state/docs.asciidoc
