+++ metricbeat/module/openmetrics/collector/_meta/data.json
@@ -11,13 +11,12 @@
     },
     "openmetrics": {
         "labels": {
-            "job": "openmetrics",
-            "listener_name": "http"
+            "job": "openmetrics"
         },
         "metrics": {
-            "net_conntrack_listener_conn_accepted_total": 3,
-            "net_conntrack_listener_conn_closed_total": 0
-        }
+            "up": 1
+        },
+        "type": "gauge"
     },
     "service": {
         "address": "127.0.0.1:55555",
