+++ metricbeat/module/prometheus/collector/_meta/data.json
@@ -11,12 +11,10 @@
     },
     "prometheus": {
         "labels": {
-            "job": "prometheus",
-            "listener_name": "http"
+            "job": "prometheus"
         },
         "metrics": {
-            "net_conntrack_listener_conn_accepted_total": 3,
-            "net_conntrack_listener_conn_closed_total": 0
+            "up": 1
         }
     },
     "service": {
